Add files via upload
This is a rocket simulator I did as a small project. It is written in Java using the java swing frame to make the applet. It uses primitive physics to calculate a rockets trajectory based on the user's input of parameters such as thrust force and mass of the rocket. It takes into account of the earth rotation as well. As of now it's pretty primitive but once I learn fluid dynamics in physics I could probably make more realistic about rocket aerodynamics etc.